added styling and form changes,modify names
