+++ public/CSS/customCSS.css
@@ -1,0 +1,79 @@
+ .highlight {
+     background-color: yellow
+ }
+ #fixed-menu
+{
+    position: fixed;
+    width: 100%;
+    height: 75px;
+    background-color: #f00;
+    padding: 10px;
+}
+.spacer
+{
+    width: 100%;
+    height: 95px;
+}
+{
+    width: 100%;
+    height: 600px;
+    background-color: #00f;
+    padding: 10px;
+}
+ #bootstrapForm{
+     width: 700px;
+     margin: 0 auto;
+ }
+html,
+body {
+  height: 100%;
+  /* The html and body elements cannot have any padding or margin. */
+}
+
+/* Wrapper for page content to push down footer */
+#wrap {
+  min-height: 100%;
+  height: 100%;
+  /* Negative indent footer by its height */
+  margin: 0 auto -82px 0;
+  /* Pad bottom by footer height */
+  padding: 0 0 80px 0;
+}
+
+/* Set the fixed height of the footer here */
+#footer {
+  height: 80px;
+  background-color: #f5f5f5;
+}
+
+
+#wrap > .center-container {
+  padding: 50px 0 0 0;
+  margin:0;
+}
+
+#footer > .container {
+  padding-top:5px
+  padding-left: 15px;
+  padding-right: 15px;
+}
+
+.center-container {
+  height:100%;
+  display: table;
+  width:100%;
+  margin:0;
+}
+
+.center-row {
+  height:50%;
+  width:100%;
+  display: table-row;
+}
+  
+.center-row > div {
+  height:100%;
+  width:50%;
+  display: table-cell;
+  border:1px solid #eee
+}
